tests/requirements.txt
coverage==4.5.1
flake8==2.5.4
mock==2.0.0
# pytest==3.4.2
pytest==3.2.4
pytest-django==3.1.2
selenium==3.10.0
tox==2.9.1
